requirements.txt
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
idna==3.11
jinja2==3.1.6
kaggle==2.0.0
kagglesdk==0.1.15
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
requests==2.32.5
setuptools==82.0.0
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typing-extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
